refactor: Using new hugginface integrations module
