+++ argilla-server/src/argilla_server/contexts/settings.py
@@ -14,10 +14,9 @@
 
 from typing import Union
 
-from argilla_server.api.schemas.v1.settings import ArgillaSettings, HuggingfaceSettings, Settings
+from argilla_server.api.schemas.v1.settings import ArgillaSettings, Settings
+from argilla_server.integrations.huggingface.spaces import HUGGINGFACE_SETTINGS, HuggingfaceSettings
 from argilla_server.settings import settings
-
-HUGGINGFACE_SETTINGS = HuggingfaceSettings()
 
 
 def get_settings() -> Settings:
